# CLI Speed improvements.