## Description

We've been trying to work on this for a while:

* Loading JSON files can take a while, so we remember their post-load state
* Concerns were made that the resulting data could be modified, so we deepcopy and return that instead
* lru_cache of 0 was actually doing nothing but logging, so sizes were adapted to be nonzero.

Before:
```
qmk find -f 'exists(split.matrix_pins)'  571.96s user 6.05s system 1137% cpu 50.815 total
```

After:
```
qmk find -f 'exists(split.matrix_pins)'  349.67s user 3.47s system 1119% cpu 31.539 total
```

## Types of Changes

- [ ] Core
- [ ] Bugfix
- [ ] New feature
- [ ] Enhancement/optimization
- [ ] Keyboard (addition or update)
- [ ] Keymap/layout/userspace (addition or update)
- [ ] Documentation

## Checklist

- [x] My code follows the code style of this project: [**C**](https://docs.qmk.fm/#/coding_conventions_c), [**Python**](https://docs.qmk.fm/#/coding_conventions_python)
- [x] I have read the [**PR Checklist** document](https://docs.qmk.fm/#/pr_checklist) and have made the appropriate changes.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**CONTRIBUTING** document](https://docs.qmk.fm/#/contributing).
- [ ] I have added tests to cover my changes.
- [x] I have tested the changes and verified that they work and don't break anything (as well as I can manage).